fix: only capture_p2p_connections once.
