+++ src/Mutations/PostMutation.php
@@ -43,7 +43,7 @@ class PostMutation extends AbstractMutation {
 				continue;
 			}
 
-			$p2p_connections_to_map = array_filter( $this->p2p_connections, [ $this, 'should_create_connection' ] );
+			$p2p_connections_to_map = Fields::get_p2p_connections();
 
 			$field_names = [];
